fix: 移除 battery_info 套件並重構 DeviceStatusService
- 移除 battery_info ^1.1.1（AGP namespace 相容性問題）
- 改用 Platform Channel 直接讀取 Android BatteryManager.EXTRA_TEMPERATURE
- 新增獨立 com.duckegg.nx4board/device_info channel（符合單一職責）
- 在 _fetch() 加入 Platform.isAndroid 保護，非 Android 平台直接跳過
- 電池溫度型別從 int? 修正為 double?（除以 10.0 的結果）

+++ lib/providers/app_provider.dart
@@ -59,7 +59,7 @@ class AppProvider extends ChangeNotifier {
   List<String> get maintenanceLogHistory => _obdService.maintenanceLogHistory;
   Stream<String> get maintenanceLogStream => _obdService.maintenanceLogStream;
   bool get isWifiConnected => _isWifiConnected;
-  int? get deviceBatteryTemp => DeviceStatusService().batteryTemperature;
+  double? get deviceBatteryTemp => DeviceStatusService().batteryTemperature;
 
   /// Initialize app - load CSV data
   Future<void> initialize() async {
